+++ packages/orq-rc/src/models/operations/updatecontact.ts
@@ -214,7 +214,7 @@ export const UpdateContactResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-11-20T15:28:09.485Z",
+    "2025-11-20T19:31:26.910Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -251,7 +251,7 @@ export const UpdateContactResponseBody$outboundSchema: z.ZodType<
   tags: z.array(z.string()).optional(),
   metadata: z.record(z.any()).optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-11-20T15:28:09.485Z"))
+  updated: z.date().default(() => new Date("2025-11-20T19:31:26.910Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {