+++ packages/orq-rc/RELEASES.md
@@ -5259,3 +5259,13 @@ Based on:
 - [typescript v4.0.0-rc.60] packages/orq-rc
 ### Releases
 - [NPM v4.0.0-rc.60] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.60 - packages/orq-rc
+
+## 2025-11-20 19:33:11
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.659.0 (2.755.9) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.0.0-rc.61] packages/orq-rc
+### Releases
+- [NPM v4.0.0-rc.61] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.61 - packages/orq-rc